requirements.txt
# Core
pyspark>=3.5.0
pandas>=2.0.0
zstandard>=0.22.0
python-dotenv>=1.0.0

# Storage
pyarrow>=14.0.0
boto3>=1.34.0

# Database
psycopg2-binary>=2.9.9

# Orchestration
apache-airflow>=2.8.0

# HTTP
requests>=2.31.0

# Testing
pytest>=7.4.0

# Optional - NLP enrichment
# spacy>=3.7.0
# transformers>=4.36.0

# Optional - Dashboard
# streamlit>=1.30.0
